lumencore/services/api/requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
PyYAML==6.0.3
